refactor: rename src/main.ts -> src/core/entry.ts
