+++ src/core/entry.ts
@@ -7,7 +7,7 @@ import { LayoutManager } from 'src/views/layout-manager'
 import { loadStyleFeatures } from 'src/rendering/style-features'
 import { codeBlockHandler } from 'src/workspace'
 import autoBind from 'auto-bind'
-import { catchInternalLinks } from './internal-links/catch-internal-links'
+import { catchInternalLinks } from 'src/internal-links/catch-internal-links'
 
 
 export default class Plugin extends ObsidianPlugin {
