+++ src/rendering/style-features.ts
@@ -1,6 +1,6 @@
 import { cssClasses } from 'src/constants'
 import { settingChanges, settingsReady } from 'src/settings/filesystem'
-import Plugin from 'src/main'
+import Plugin from 'src/core/entry'
 import Callbag from 'src/utilities/callbag'
 import { layoutReady } from 'src/utilities/layout-ready'
 import { globalStyle, toggleBodyClass, settingTriggers as t, themeChange } from './style-tools'
